requirements.txt
pandas
selenium
boto3
scikit-learn
matplotlib
seaborn
python-dotenv